refactor: organize entity folders
