+++ src/main/java/dev/galacticraft/mod/client/render/entity/EvolvedEvokerEntityRenderer.java
@@ -24,7 +24,7 @@ package dev.galacticraft.mod.client.render.entity;
 
 import dev.galacticraft.mod.client.render.entity.feature.EvolvedIllagerMaskRenderLayer;
 import dev.galacticraft.mod.client.render.entity.feature.OxygenTanksRenderLayer;
-import dev.galacticraft.mod.content.entity.EvolvedEvokerEntity;
+import dev.galacticraft.mod.content.entity.evolved.EvolvedEvokerEntity;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.EvokerRenderer;
