+++ src/main/java/dev/galacticraft/mod/client/render/entity/GreyEntityRenderer.java
@@ -26,8 +26,8 @@ import dev.galacticraft.mod.Constant;
 import dev.galacticraft.mod.client.model.entity.ArchGreyEntityModel;
 import dev.galacticraft.mod.client.model.entity.GreyEntityModel;
 import dev.galacticraft.mod.client.render.entity.model.GCEntityModelLayer;
-import dev.galacticraft.mod.content.entity.ArchGreyEntity;
-import dev.galacticraft.mod.content.entity.GreyEntity;
+import dev.galacticraft.mod.content.entity.moon.ArchGreyEntity;
+import dev.galacticraft.mod.content.entity.moon.GreyEntity;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.MobRenderer;
