+++ src/main/java/dev/galacticraft/mod/client/render/entity/OliGrubEntityRenderer.java
@@ -25,7 +25,7 @@ package dev.galacticraft.mod.client.render.entity;
 import dev.galacticraft.mod.Constant;
 import dev.galacticraft.mod.client.model.entity.OliGrubEntityModel;
 import dev.galacticraft.mod.client.render.entity.model.GCEntityModelLayer;
-import dev.galacticraft.mod.content.entity.OliGrubEntity;
+import dev.galacticraft.mod.content.entity.moon.OliGrubEntity;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.MobRenderer;
 import net.minecraft.resources.ResourceLocation;
