+++ src/main/java/dev/galacticraft/mod/content/block/entity/DungeonSpawnerBlockEntity.java
@@ -23,10 +23,10 @@
 package dev.galacticraft.mod.content.block.entity;
 
 import dev.galacticraft.mod.content.GCBlockEntityTypes;
-import dev.galacticraft.mod.content.entity.EvolvedCreeperEntity;
-import dev.galacticraft.mod.content.entity.EvolvedSkeletonEntity;
-import dev.galacticraft.mod.content.entity.EvolvedSpiderEntity;
-import dev.galacticraft.mod.content.entity.EvolvedZombieEntity;
+import dev.galacticraft.mod.content.entity.evolved.EvolvedCreeperEntity;
+import dev.galacticraft.mod.content.entity.evolved.EvolvedSkeletonEntity;
+import dev.galacticraft.mod.content.entity.evolved.EvolvedSpiderEntity;
+import dev.galacticraft.mod.content.entity.evolved.EvolvedZombieEntity;
 import dev.galacticraft.mod.content.entity.boss.AbstractBossEntity;
 import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
