+++ src/main/java/dev/galacticraft/mod/content/entity/evolved/EvolvedEndermanEntity.java
@@ -20,34 +20,30 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.entity;
+package dev.galacticraft.mod.content.entity.evolved;
 
-import dev.galacticraft.api.entity.attribute.GcApiEntityAttributes;
+import dev.galacticraft.mod.Constant;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.FlyingMob;
-import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
-import net.minecraft.world.entity.ai.attributes.Attributes;
-import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal;
-import net.minecraft.world.entity.monster.Enemy;
-import net.minecraft.world.entity.monster.Monster;
+import net.minecraft.world.entity.monster.EnderMan;
 import net.minecraft.world.level.Level;
 
-public class CometCubeEntity extends FlyingMob implements Enemy {
-    public CometCubeEntity(EntityType<? extends FlyingMob> entityType, Level level) {
-        super(entityType, level);
+public class EvolvedEndermanEntity extends EnderMan {
+    public EvolvedEndermanEntity(EntityType<? extends EvolvedEndermanEntity> entityType, Level world) {
+        super(entityType, world);
     }
 
     @Override
-    protected void registerGoals() {
-        this.goalSelector.addGoal(0, new RandomLookAroundGoal(this));
+    public boolean galacticraft$hasMask() {
+        return true;
     }
 
-    public static AttributeSupplier.Builder createAttributes() {
-        return Monster.createMonsterAttributes()
-                .add(Attributes.MOVEMENT_SPEED, 0.35F)
-                .add(Attributes.FOLLOW_RANGE, 12.0)
-                .add(Attributes.MAX_HEALTH, 24.0)
-                .add(Attributes.ATTACK_DAMAGE, 5.0)
-                .add(GcApiEntityAttributes.CAN_BREATHE_IN_SPACE, 1.0D);
+    @Override
+    public boolean galacticraft$hasGear() {
+        return true;
+    }
+
+    @Override
+    public String galacticraft$tankSize(int i) {
+        return Constant.Item.MEDIUM_OXYGEN_TANK;
     }
 }
