+++ src/main/java/dev/galacticraft/mod/content/entity/evolved/EvolvedSkeletonEntity.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.entity;
+package dev.galacticraft.mod.content.entity.evolved;
 
 import dev.galacticraft.mod.Constant;
 import dev.galacticraft.mod.mixin.AbstractSkeletonEntityAccessor;
