+++ src/main/java/dev/galacticraft/mod/content/entity/moon/GreyEntity.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.entity;
+package dev.galacticraft.mod.content.entity.moon;
 
 import dev.galacticraft.api.entity.attribute.GcApiEntityAttributes;
 import dev.galacticraft.mod.content.entity.goals.FollowPlayerGoal;
