+++ src/main/java/dev/galacticraft/mod/content/entity/moon/OliGrubEntity.java
@@ -20,24 +20,43 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.entity;
+package dev.galacticraft.mod.content.entity.moon;
 
 import dev.galacticraft.api.entity.attribute.GcApiEntityAttributes;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.AgeableMob;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.AvoidEntityGoal;
 import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal;
+import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.monster.Monster;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.Nullable;
 
-public class GazerEntity extends Monster {
-    public GazerEntity(EntityType<? extends Monster> entityType, Level level) {
+public class OliGrubEntity extends Animal {
+    public OliGrubEntity(EntityType<? extends Animal> entityType, Level level) {
         super(entityType, level);
     }
 
     @Override
+    public boolean isFood(ItemStack stack) {
+        return false;
+    }
+
+    @Nullable
+    @Override
+    public AgeableMob getBreedOffspring(ServerLevel serverLevel, AgeableMob ageableMob) {
+        return null;
+    }
+
+    @Override
     protected void registerGoals() {
-        this.goalSelector.addGoal(0, new RandomLookAroundGoal(this));
+        this.goalSelector.addGoal(0, new AvoidEntityGoal<>(this, Player.class, 16.0F, 0.8, 1.33));
+        this.goalSelector.addGoal(1, new RandomLookAroundGoal(this));
     }
 
     public static AttributeSupplier.Builder createAttributes() {
